case-insensitive sorting on map
