+++ statusboard/src/utils/utils.ts
@@ -1,4 +1,7 @@
 import { Bounds, latLng } from 'leaflet';
+
+import { SortByFn, Row, IdType } from 'react-table';
+
 import { Brigade, Project } from './types';
 
 export type ActiveThresholdsKeys = 'all time' | 'year' | 'month' | 'week';
@@ -164,3 +167,51 @@ export function getTopicsFromProjects(projects?: Project[]) {
     .sort((a, b) => b[1] - a[1])
     .map((topicAndCount) => topicAndCount[0]);
 }
+
+function periodToNumber(period: string) {
+  if (period === 'week') {
+    return 0;
+  }
+  if (period === 'month') {
+    return 1;
+  }
+  if (period === 'year') {
+    return 2;
+  }
+  if (period === 'over_a_year') {
+    return 3;
+  }
+  return 99;
+}
+
+export const customStringSort: SortByFn<Project> = (
+  rowA: Row<Project>,
+  rowB: Row<Project>,
+  id: IdType<Project>,
+  desc?: boolean
+): number => {
+  const rowAValue: string = rowA.values[id] ?? '';
+  const rowBValue: string = rowB.values[id] ?? '';
+
+  const valueA: string = String(rowAValue).toLowerCase();
+  const valueB: string = String(rowBValue).toLowerCase();
+  if (desc) {
+    return valueA.localeCompare(valueB) > 0 ? 1 : -1;
+  }
+  return valueB.localeCompare(valueA) > 0 ? -1 : 1;
+};
+
+export const lastPushSort: SortByFn<Project> = (
+  rowA: Row<Project>,
+  rowB: Row<Project>,
+  id: IdType<Project>,
+  desc?: boolean
+): number => {
+  // need to convert week, month to 1, 2, ..
+  const d1: number = periodToNumber(rowA.values[id]);
+  const d2: number = periodToNumber(rowB.values[id]);
+  if (desc) {
+    return d1 - d2 > 0 ? 1 : -1;
+  }
+  return d2 - d1 > 0 ? -1 : 1;
+};
